pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "flat-py"
version = "0.1.dev0"
description = "Runtime type checker for formal language types in Python"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "parsy>=2.2.0",
    "isla-solver>=1.14.4",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.18.0",
]

[tool.mypy]
python_version = 3.10
exclude = [
    'examples'
]